[FIX] #13 - 등급 변경 오류 수정
