+++ app.js
@@ -1,9 +1,9 @@
-  var createError = require('http-errors');
-  var express = require('express');
-  var path = require('path');
-  var cookieParser = require('cookie-parser');
-  var logger = require('morgan');
-  var session = require('express-session');
+var createError = require('http-errors');
+var express = require('express');
+var path = require('path');
+var cookieParser = require('cookie-parser');
+var logger = require('morgan');
+var session = require('express-session');
 
 
 var homeRouter = require('./routes/home');
@@ -20,28 +20,25 @@ var Freeboard = require('./routes/Freeboard');
 var Reportboard = require('./routes/Reportboard');
 
 
-  var app = express();
+var app = express();
 
-  // view engine setup
-  app.set('views', path.join(__dirname, 'views'));
-  app.set('view engine', 'ejs');
+// view engine setup
+app.set('views', path.join(__dirname, 'views'));
+app.set('view engine', 'ejs');
 
-  // 세션 미들웨어 구성
-  app.use(session({
-    secret: 'dbproject2023', // 비밀 키 설정
-    resave: false,
-    saveUninitialized: true,
-    cookie: { secure: false } // HTTPS를 사용하지 않는 경우 false로 설정
-  }));
+// 세션 미들웨어 구성
+app.use(session({
+  secret: 'dbproject2023', // 비밀 키 설정
+  resave: false,
+  saveUninitialized: true,
+  cookie: { secure: false } // HTTPS를 사용하지 않는 경우 false로 설정
+}));
 
-  app.use(logger('dev'));
-  app.use(express.json());
-  app.use(express.urlencoded({ extended: false }));
-  app.use(cookieParser());
-  app.use(express.static(path.join(__dirname, 'public')));
-
-
- 
+app.use(logger('dev'));
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', homeRouter);
 app.use('/users', usersRouter);
@@ -55,33 +52,34 @@ app.use('/logout', logoutRouter);
 app.use('/Freeboard', Freeboard);
 app.use('/Reportboard', Reportboard);
 
+const updateUserRole = require('./public/viewscripts/userRole');
+
 app.post('/changeRole', (req, res) => {
   const userId = req.body.id;
   const newRole = req.body.role;
 
   updateUserRole(userId, newRole)
-      .then(message => res.send({ message }))
-      .catch(error => res.status(500).send({ error }));
+    .then(message => res.send({ message }))
+    .catch(error => res.status(500).send({ error }));
 });
+
 app.use('/myinfo', myinfoRouter);
 
+// catch 404 and forward to error handler
+app.use(function (req, res, next) {
+  next(createError(404));
+});
 
+// error handler
+app.use(function (err, req, res, next) {
+  // set locals, only providing error in development
+  res.locals.message = err.message;
+  res.locals.error = req.app.get('env') === 'development' ? err : {};
+  console.error(err.stack); // 에러 스택 트레이스 로깅
+  // render the error page
+  res.status(err.status || 500);
+  res.render('error');
+});
 
-  // catch 404 and forward to error handler
-  app.use(function(req, res, next) {
-    next(createError(404));
-  });
-
-  // error handler
-  app.use(function(err, req, res, next) {
-    // set locals, only providing error in development
-    res.locals.message = err.message;
-    res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-    // render the error page
-    res.status(err.status || 500);
-    res.render('error');
-  });
-
-  module.exports = app;
+module.exports = app;
