+++ views/userInfo.ejs
@@ -2,7 +2,9 @@
 <html>
 
 <head>
-  <title><%= title %></title>
+  <title>
+    <%= title %>
+  </title>
   <link rel='stylesheet' href='/stylesheets/manage.css' />
 </head>
 
@@ -10,8 +12,12 @@
   <div class="menu-logo" id="hiddenMenu">
     <img src="/images/menu.png" alt="메뉴 로고">
   </div>
-  <h1><%= title %></h1>
-  <h2><%= subtitle %></h2>
+  <h1>
+    <%= title %>
+  </h1>
+  <h2>
+    <%= subtitle %>
+  </h2>
   <table>
     <thead>
       <tr>
@@ -28,22 +34,40 @@
       </tr>
     </thead>
     <tbody>
-      <% for(var i = 0; i < rows.length; i++) { %>
-      <tr>
-        <td><%= rows[i].Id %></td>
-        <td><%= rows[i].Email %></td>
-        <td><%= rows[i].Name %></td>
-        <td><%= rows[i].PhoneNum %></td>
-        <td><%= rows[i].Address %></td>
-        <td><%= rows[i].Gender %></td>
-        <td><%= rows[i].Age %></td>
-        <td><%= rows[i].Role %></td>
-        <td><%= rows[i].UStatus === 1 ? '가입' : '탈퇴' %></td>
-        <td>
-          <button onclick="changeRole('<%= rows[i].Id %>', '<%= rows[i].Role %>')">변경</button>
-        </td>
-      </tr>
-      <% } %>
+      <% for(var i=0; i < rows.length; i++) { %>
+        <tr>
+          <td>
+            <%= rows[i].Id %>
+          </td>
+          <td>
+            <%= rows[i].Email %>
+          </td>
+          <td>
+            <%= rows[i].Name %>
+          </td>
+          <td>
+            <%= rows[i].PhoneNum %>
+          </td>
+          <td>
+            <%= rows[i].Address %>
+          </td>
+          <td>
+            <%= rows[i].Gender %>
+          </td>
+          <td>
+            <%= rows[i].Age %>
+          </td>
+          <td>
+            <%= rows[i].Role %>
+          </td>
+          <td>
+            <%= rows[i].UStatus===1 ? '가입' : '탈퇴' %>
+          </td>
+          <td>
+            <button onclick="changeRole('<%= rows[i].Id %>', '<%= rows[i].Role %>')">변경</button>
+          </td>
+        </tr>
+        <% } %>
     </tbody>
   </table>
   <div class="menu-area">
@@ -109,33 +133,33 @@
       }
 
       if (confirm(message)) {
-      // 서버에 변경 요청을 보냄
-      fetch('/changeRole', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ id: userId, role: newRole })
-      })
-      .then(response => {
-        if(response.ok) {
-          return response.json();
-        }
-        throw new Error('Request failed.');
-      })
-      .then(data => {
-        console.log(data);
-        // 성공 메시지나 변경된 상태를 화면에 표시하는 로직을 추가할 수 있습니다.
-      })
-      .catch(error => {
-        console.error('Error:', error);
-      });
+        // 서버에 변경 요청을 보냄
+        fetch('/changeRole', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({ id: userId, role: newRole })
+        })
+          .then(response => {
+            if (response.ok) {
+              return response.json();
+            }
+            throw new Error('Request failed.');
+          })
+          .then(data => {
+            console.log(data);
+            // 성공 메시지나 변경된 상태를 화면에 표시하는 로직을 추가할 수 있습니다.
+          })
+          .catch(error => {
+            console.error('Error:', error);
+          });
 
-      console.log("등급 변경 요청:", userId, "새 등급:", newRole);
-    } else {
-      // 취소한 경우
-      console.log("등급 변경 취소됨");
-    }
+        console.log("등급 변경 요청:", userId, "새 등급:", newRole);
+      } else {
+        // 취소한 경우
+        console.log("등급 변경 취소됨");
+      }
     }
   </script>
   <script src="/viewscripts/logout.js"></script>
